Embed questions using GResource
This improves portability drastically because I don't need to think
about a packaging data files. I just distribute the binary.

Obviously GResource isn't appropriate for everything, but for a game
like this it's perfect.

Unfortunately this breaks the handy ./llc.vala workflow, because
this requires compiling a .c file into the binary, and `vala` doesn't
understand how to do that (even though its twin `valac` is perfectly
aware of how).